requirements.txt
requests
pygetwindow
guessit
python-dotenv
pywin32
psutil
PySimpleGUIQt